==> 00_DL_preview.ipynb <==
# Deep Learning

- Artificial neural network(ANN) = Deep neural network(DNN) = Deep learning(DL)
- 하나하나..곱셈 엄청 많이 일어남(합성곱) >> GPU필요(많은 연산 하려면!)

- 이 '문제'를 잘 해결하기 위해서는 '어떤 방법', '어떤 모델'을 사용하는게 좋을까?? >> 모든게 다 이 질문에서 시작!!

- 정형데이터는 머신러닝 good
- 비정형데이터는 딥러닝!
- 이미지, 비디오, 텍스트 데이터 등.. 세상에 비정형데이터 많음! 머신러닝으로 해결하기 어려워 >> 그래서 딥러닝!!
- 컴퓨터비전, 자연어 처리 + 추천시스템

- GAN: 생성모델.. 위조지폐범..
- 저작권이 없는 이미지.. 만들수 있음.. 이미지장사 할 수도 있음.. / 지금은 검색해서 쓰지만..흠
- 딥러닝 가장 처음 나온게 2010년! 제대로된 연구는 16년에 시작함

- 퍼셉트론: 뭔가들어가서 뭔가 나옴(ex. 맞으면 아파..'아')
- 활성화 함수(인풋데이터, 가중치 받아서 이걸 어떻게 내보내줄지 결정해주는 함수.. 얘가 출력해주는건 다음 layer의 input이 됨. (입력~ 출력, 이거 하나가 퍼셉트론 하나임.)
- 동그라미 하나가 퍼셉트론임!! 하나의 레이어의 여러 퍼셉트론이 있음.
- ANN을 많은 층으로 깊게 쌓은 모델이 DNN = DL

활성화 함수
- Relu: 0이하면 버리고, 이상이면 그대로

머신러닝vs딥러닝
- feature engineering(데이터 전처리): 머신러닝은 이것때문에 시간이 오래걸림, 딥러닝은 이거x

- 딥러닝은 해석이 불가능하다.. 이런게 어렵다..(머신러닝은 딥러닝보다는 해석할 수 있는 여지가 더 o ex.importance__

Deep Learning Terms
- backpropagation: weight가 싹바뀌고, 바뀌고..함!
- optimizer: 히말라야 세르파 같은 역할.. cost가 가장 작은 지점으로 데리러 가는..cost function오르게 하는..흠.

활성화 함수
- sigmoid, leaky relu, relu, tanh 이 4개는 많이 쓰임!

Mini Gradient Descent
- Gradient Descent 방식 중 이걸 제일 많이 활용함!

Optimizer 종류
- RMSProopo Optimizer, Adam Optimizer 를 제일 많이씀!
- 그림 >> 속도, 얼마나 튀는지

Cost Function
- 어떤 문제냐에 따라서 사용해야 할 cost function이 다름!
- MSE, Binary Cross Entropy(0과 1로 분류), Categorical Cross Entropy(one-hot-vector), Sparse Categorical Crossentropy(integer) .... 

요소들 바꿔보면서 어떻게 동작하는지 보기!
- http://playground.tensorflow.org/#activation=tanh&batchSize=10&dataset=circle&regDataset=[…]se&problem=classification&initZero=false&hideText=false

- 오차계산하고, 옵티마이저가 틀렸으니까 바꿔(weight) >> backpropagation (이전보다 덜 틀리는 방향으로 수정됨, 중요한건 더 중요하게 되겠지?!!)
- batch로 나누는 이유? 속도향상?!?!(batch = 데이터를 한 뭉탱이로 묶는 것)(batch size 256 >> 256개 데이터를 묶는다) 비치개수 만큼이 가중치..붙는거!
- epoch: 전체 데이터를 활용하여 학습하는 횟수

Overview PPT
- 이 데이터들로 할 수 있는것
- 지도학습, 회귀.. 보험료..흠.. >> 지도학습 regression

- 데이터가 만개 정도면.. 딥러닝 쓰기 좋지 않을까.. >> 암튼 데이터가 많을수록 좋음!

용어
- backpropagation: 주로 업데이트 되는게 weight임!
- optimizer: 오차를 내려가게 하는 방식도 굉장히 다양함! 
- layer중간에 많이 쓰는게 그 4개, 거의 출력할때만 쓰는게 softmax function!!(총합 1되게끔)

Global minimum, Local minimum
- 우리가 찾은 최솟값이 실제 최솟값이 아닐 수 있음!
- 전체 데이터를 활용할지, 데이터를 묶어 사용하던지 하는 방법으로 해결하려고 함
- Mini gradient descent >> 거의 얘밖에 안씀!
- BGD(정직, shooting이 없음, local 에 빠지면 나오기 힘듬, 한번에 보는 느낌)
- SGD(shooting 있음 > global 찾을 가능성 있음, 하나하나 봐서 오래걸림
- 속도도 비용임!(좋은 컴퓨터 쓰는 이유가 .. 이거!)
