==> 08_RNN 기초.ipynb <==
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import SimpleRNN

model = Sequential()
model.add(SimpleRNN(3, input_shape = (2, 10)))
# model.add(SimpleRNN(hidden_units, input_shape = (timesteps, input_dim)))
model.summary()

model = Sequential()
# timesteps = input_length
model.add(SimpleRNN(3, input_length = 2, input_dim = 10))
model.summary()

### USE batch

model = Sequential()
# model.add(SimpleRNN(hidden unit,
# batch_input_shape = (batch_size , timesteps(input_length), input_dim))) 
model.add(SimpleRNN(3, batch_input_shape = (8,2,10)))
model.summary()

### Use batch / resturn_sequence = True

model = Sequential()
# model.add(SimpleRNN(hidden unit,
# batch_input_shape = (batch_size , timesteps(input_length), input_dim))) 
model.add(SimpleRNN(3, batch_input_shape = (8,2,10), return_sequences = True))
model.summary()

### 깊은 순환신경망

model = Sequential()
# timesteps = input_length
model.add(SimpleRNN(3, input_length = 10, input_dim = 5, return_sequences=True))
model.add(SimpleRNN(3, return_sequences=True))
model.summary()

# Simple RNN(use Data)

import numpy as np
import tensorflow as tf
from tensorflow.keras.layers import SimpleRNN, LSTM

# time step 4, input_dim = 5, batch_size = 1
train_X = [[
    [0.1, 4.2, 1.5, 1.1, 2.8], 
    [1.0, 3.1, 2.5, 0.7, 1.1],
    [0.3, 2.1, 1.5, 2.1, 0.1],
    [2.2, 1.4, 0.5, 0.9, 1.1]
    ]]
train_X = np.array(train_X, dtype=np.float32)
print(train_X.shape)

### hidden_unit = 3

rnn = SimpleRNN(3)
hidden_states = rnn(train_X)
print(f'input shape : {train_X.shape}')
print(train_X)
print(f'hidden(output) shape : {hidden_states.shape}')
print(hidden_states)

### hidden_unit = 3, return_sequence = True

rnn = SimpleRNN(3, return_sequences = True)
hidden_states = rnn(train_X)
print(f'input shape : {train_X.shape}')
print(train_X)
print(hidden_states)

### hidden_unit = 3, return_sequence = True, return_state = True
- return_state = True인 경우에는 return_sequence의 True/False와 관계없이 last_states는 마지막 hidden state를 출력

rnn = SimpleRNN(3, return_sequences = True, return_state = True)
hidden_states, last_states = rnn(train_X)
print('-----input-----')
print(f'input shape : {train_X.shape}')
print(train_X)
print('-----hidden-----')
print(hidden_states)
print('-----last-----')
print(last_states)

### hidden_unit = 3, return_sequence = False, return_state = True

rnn = SimpleRNN(3, return_sequences = False, return_state = True)
hidden_states, last_states = rnn(train_X)
print('-----input-----')
print(f'input shape : {train_X.shape}')
print(train_X)
print('-----hidden-----')
print(hidden_states)
print('-----last-----')
print(last_states)

# LSTM

### return_sequences = False, return_state = True

lstm = LSTM(3, return_sequences = False, return_state = True)
hidden_states, last_states, last_cell_state = lstm(train_X)
print('-----input-----')
print(f'input shape : {train_X.shape}')
print(train_X)
print('-----hidden-----')
print(hidden_states)
print('-----last-----')
print(last_states)
print('-----last cell-----')
print(last_cell_state)

### return_sequences = True, return_state = True

lstm = LSTM(3, return_sequences = True, return_state = True)
hidden_states, last_states, last_cell_state = lstm(train_X)
print('-----input-----')
print(f'input shape : {train_X.shape}')
print(train_X)
print('-----hidden-----')
print(hidden_states)
print('-----last-----')
print(last_states)
print('-----last cell-----')
print(last_cell_state)
