==> 15_used word2vec modeling.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

# word2vec을 활용한 리뷰 감정 예측 모델

import pandas as pd
from gensim.models import KeyedVectors
# 전처리가 완료된 데이터
tokenized_data = pd.read_pickle(r'/content/drive/MyDrive/8_DL/tokenized_data.pkl')
label_data = pd.read_pickle(r'/content/drive/MyDrive/8_DL/labels.pkl')
loaded_model = KeyedVectors.load_word2vec_format('/content/drive/MyDrive/8_DL/Word2Vec')

# # 이렇게 해도 됨

# from google.colab import drive
# drive.mount('/content/drive')

# # 개인의 구글 드라이브 폴더 path를 넣어줘야 합니다.
# dir_path = '/content/drive/MyDrive/lecture/진행 중/서울산업진흥원(SBA)/02. 딥러닝/05. NLP/data'

# import os
# os.chdir(dir_path) # 현재 경로 수정

# import pandas as pd
# from gensim.models import KeyedVectors
# # 전처리가 완료된 데이터
# tokenized_data = pd.read_pickle(r'tokenized_data.pkl')
# label_data = pd.read_pickle(r'labels.pkl')
# loaded_model = KeyedVectors.load_word2vec_format('Word2Vec')

tokenized_data[0] # 얘들을 평균의 벡터로 보겠다

# sentence는 여러개의 word로 되어 있어서 하나의 sentence vector를 word 벡터들의 평균으로 지정
loaded_model[tokenized_data[0]].mean(axis = 0)

from tqdm import tqdm
import numpy as np

data_vectors = []
embedding_size = 100
for tokened_sentence in tqdm(tokenized_data): # 토큰화된 센텐스가 들어옴
  try:
    senetence_vector = loaded_model[tokened_sentence].mean(axis = 0)
    data_vectors.append(senetence_vector)
  except:
    zeros_vector = np.zeros(embedding_size)
    data_vectors.append(zeros_vector)

vector_data = np.array(data_vectors) # 데이터가 숫자로 바뀜. 이제 똑같이 모델 생성, 학습해주면됨!

vector_data.shape

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(vector_data, label_data, test_size = 0.25)
print(X_train.shape)
print(X_test.shape)
print(y_train.shape)
print(y_test.shape)

from sklearn.linear_model import LogisticRegression # 분류 모델
lr_model = LogisticRegression()
lr_model.fit(X_train, y_train)

lr_model.score(X_train, y_train)

lr_model.score(X_test, y_test)

# randomforest

from sklearn.ensemble import RandomForestClassifier
rf = RandomForestClassifier(random_state=0, n_estimators=100) # max_depth 안주니까 성능올라감
rf.fit(X_train, y_train)

rf.get_params()

rf.score(X_train, y_train)

rf.score(X_test, y_test)

!pip install konlpy

# 전처리 > 모델 활용해서 벡터로 바꿔주기 

from tqdm import tqdm
from konlpy.tag import Okt
import re

def predict_text(text):
  # 데이터 전처리
  stopwords = ['의', '를', '와', '과', '으로', '것', '잘', '되', '주', '보', '수',
              '좀', '이', '있', '은', '한', '아', '도', '에', '하', '는', '자',
              '하다', '가', '들', '걍', '등']

  text = re.sub('[^ㄱ-ㅎㅏ-ㅣ가-힣 ]','', text)

  okt = Okt()
  tokenized_data = []

  tokenized_sentence = okt.morphs(text, stem = True)
  stopwords_removed_sentences = [token for token in tokenized_sentence if not token in stopwords]
  tokenized_data.append(stopwords_removed_sentences)
  
  # 데이터 벡터화
  data_vectors = []
  embedding_size = 100
  for tokened_sentence in tqdm(tokenized_data):
    try:
      sentence_vector = loaded_model[tokened_sentence].mean(axis = 0)
      data_vectors.append(sentence_vector)
    except:
      zeros_vector = np.zeros(embedding_size)
      data_vectors.append(zeros_vector)
  data_vectors = np.array(data_vectors)
  
  # 예측 및 결과 정리
  prediction = rf.predict(data_vectors)[0]
  if prediction == 1:
    answer = '긍정 리뷰입니다.'
  else:
    answer = '부정 리뷰입니다.'

  return answer

predict_text('노잼 노잼 노잼')

### 예측하기(강의 전)

!pip install konlpy
