==> 16_tensorflow_벡터화.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

# 개인의 구글 드라이브 폴더 path를 넣어줘야 합니다.
dir_path = '/content/drive/MyDrive/8_DL'

import os
os.chdir(dir_path) # 현재 경로 수정

import pandas as pd
# 전처리가 완료된 데이터
tokenized_data = pd.read_pickle(r'tokenized_data.pkl')
label_data = pd.read_pickle(r'labels.pkl')

import numpy as np
train_data = np.array(tokenized_data)

from sklearn.model_selection import train_test_split
tokenized_X_train, tokenized_X_test, y_train, y_test = train_test_split(train_data, label_data, test_size = 0.25)
print(tokenized_X_train.shape)
print(tokenized_X_test.shape)
print(y_train.shape)
print(y_test.shape)

from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.preprocessing.text import Tokenizer
tokenizer = Tokenizer()
tokenizer.fit_on_texts(tokenized_X_train)
train_sequences = tokenizer.texts_to_sequences(tokenized_X_train)
test_sequences = tokenizer.texts_to_sequences(tokenized_X_test)


word_vocab = tokenizer.word_index # 단어 : integer
word_vocab['<PAD>'] = 0 # 패드해줄 vocab 추가 

# vocab size 가 필요함. 몇개있는지.
import json
vocab = {} # 딕셔너리 형태니까 JSON으로
vocab['vocab'] = word_vocab
vocab['vocab_size'] = len(word_vocab)

json.dump(vocab, open(f'{dir_path}/config.json', 'w', ensure_ascii=False))

len(word_vocab)

SEQUENCE_MAX_LEN = 8
train_inputs = pad_sequences(train_sequences, maxlen = SEQUENCE_MAX_LEN, padding = 'post')
test_inputs = pad_sequences(test_sequences, maxlen = SEQUENCE_MAX_LEN, padding = 'post')

print(train_inputs.shape)
print(test_inputs.shape)

import pickle
 
# Open a file and use dump()
with open(f'{dir_path}/tf_train_inputs.pkl', 'wb') as file: 
  # A new file will be created
  pickle.dump(train_inputs, file) 
with open(f'{dir_path}/tf_test_inputs.pkl', 'wb') as file: 
  pickle.dump(test_inputs, file) 
with open(f'{dir_path}/tf_train_labels.pkl', 'wb') as file: 
  pickle.dump(y_train, file) 
with open(f'{dir_path}/tf_test_labels.pkl', 'wb') as file: 
  pickle.dump(y_test, file)
