==> 18_tensorflow(1D CNN).ipynb <==
# layer

## GlobalMaxPooling1D

import tensorflow as tf
from tensorflow.keras.layers import MaxPooling1D,GlobalMaxPooling1D
x = tf.constant([[1., 2., 3.],
                 [4., 5., 6.],
                 [7., 8., 9.]])
print(x.shape)
print(x)

x = tf.reshape(x, [3, 3, 1])
print(x.shape)
print(x)

max_pool_1d = GlobalMaxPooling1D()
max_pool_1d(x)

## MaxPooling1D

x = tf.constant([1., 2., 3., 4., 5.])
x = tf.reshape(x, [1, 5, 1])
print(x)

max_pool_1d = MaxPooling1D(pool_size=2,
    strides=1, padding='valid')
max_pool_1d(x)

# Modeling

from google.colab import drive
drive.mount('/content/drive')

import os
import pandas as pd
import json
# 개인의 구글 드라이브 폴더 path를 넣어줘야 합니다.
dir_path = '/content/drive/MyDrive/lecture/진행 중/서울산업진흥원(SBA)/02. 딥러닝/05. NLP/data'
os.chdir(dir_path) # 현재 경로 수정

# 전처리 & 벡터화가 완료된 데이터
X_train = pd.read_pickle(f'{dir_path}/tf_train_inputs.pkl')
X_text = pd.read_pickle(f'{dir_path}/tf_test_inputs.pkl')
y_train = pd.read_pickle(f'{dir_path}/tf_train_labels.pkl')
y_test = pd.read_pickle(f'{dir_path}/tf_test_labels.pkl')

config = json.load(open(f'{dir_path}/config.json', 'r'))
vocab = config['vocab']
vocab_size = config['vocab_size']

from keras.api._v2.keras import activations
from tensorflow.keras.layers import Embedding, Dense, Conv1D, Dropout, MaxPooling1D, GlobalMaxPooling1D
from tensorflow.keras.models import Sequential, load_model
from tensorflow.keras.callbacks import EarlyStopping, ModelCheckpoint

embedding_dim = 100
model = Sequential()
model.add(Embedding(vocab_size, embedding_dim))
model.add(Dropout(0.3))
model.add(Conv1D(32, 5, padding = 'valid', activation = 'relu'))
model.add(GlobalMaxPooling1D())
model.add(Dropout(0.3))
model.add(Dense(1, activation = 'sigmoid'))
es = EarlyStopping(monitor = 'val_loss', mode = 'min', verbose = 1, patience = 4)
mc = ModelCheckpoint('best_1dcnn.h5', monitor = 'val_acc', mode = 'max', verbose = 1, save_best_only = True)

model.compile(optimizer = 'rmsprop', loss = 'binary_crossentropy', metrics = ['acc'])
history = model.fit(X_train, y_train,
                    epochs = 30, callbacks = [es, mc],
                    batch_size = 64,validation_split = 0.2)

import pickle
# load the tokenizer from a file
with open('tokenizer.pickle', 'rb') as handle:
    tokenizer = pickle.load(handle)

!pip install konlpy

load_model = load_model(f'{dir_path}/best_1dcnn.h5')

# text를 넣어 확률을 활용하여 리뷰를 예측하는 함수
text = '진짜 너무 재밌다.'
from tqdm import tqdm
from konlpy.tag import Okt
import re
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.preprocessing.text import Tokenizer

def predict_text(text):
  # 데이터 전처리
  stopwords = ['의', '를', '와', '과', '으로', '것', '잘', '되', '주', '보', '수',
              '좀', '이', '있', '은', '한', '아', '도', '에', '하', '는', '자',
              '하다', '가', '들', '걍', '등']

  text = re.sub('[^ㄱ-ㅎㅏ-ㅣ가-힣 ]','', text)

  okt = Okt()
  tokenized_data = []

  tokenized_sentence = okt.morphs(text, stem = True)
  stopwords_removed_sentences = [token for token in tokenized_sentence if not token in stopwords]
  tokenized_data.append(stopwords_removed_sentences)

  # 벡터화
  sequences = tokenizer.texts_to_sequences(tokenized_data)

  SEQUENCE_MAX_LEN = 8
  inputs = pad_sequences(sequences , maxlen = SEQUENCE_MAX_LEN, padding = 'post')
  score = load_model.predict(inputs)[0]
  if score > 0.7:
    print(f'{round(score[0] * 100, 2)}% 긍정 리뷰 입니다.')
  else:
    print(f'{round(score[0] * 100, 2)}% 부정 리뷰 입니다.')
  return score


predict_text('와 진짜 꿀잼')
